update verticalresourcecardlist to actionList
