+++ src/components/FeatureFeed/FeatureFeedComponentMap.js
@@ -4,7 +4,8 @@ import {
   HorizontalMediaListFeature,
   VerticalCardListFeature,
   ButtonFeature,
-  VerticalResourceCardListFeature,
+  ActionListFeature,
+  ChipListFeature,
 } from './Features';
 
 // Maps a graphql "__typename" to a Component to render it.
@@ -14,7 +15,8 @@ const FeatureFeedComponentMap = {
   HorizontalMediaListFeature,
   VerticalCardListFeature,
   ButtonFeature,
-  VerticalResourceCardListFeature,
+  ActionListFeature,
+  ChipListFeature,
 };
 
 export default FeatureFeedComponentMap;
